+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Examine all destination slots - kylev
 * Armory Cabinet providers for Jade and TOP - Christofmeg
 * Fixed being able to extract upgrades when going over the possible limit closes #406 blame @Matyrobbrt
+* Added extra checks for the reconnection of the Controller Extension, closes #395
 
 # VERSION 1.4.2
 * Fix upgrades not being insertable/extractable in compacting drawers by Matyrobbrt